Wow, the VBO currently looks great! Just kidding, I have to do a ton of
optimization on the code and make it completely unreadable so that it
has the highest performance java will allow :P. Worst thing is removing
the for loops because of the additional stupid increment after
determining to exit the for loop. This isn't much of a big deal, but I
don't want any extra calls to the CPU with the most important rendering
portion of the code.