I might have to change how the images are loaded because I just realized
a mistake in the process. Anyways, I added a texturearray object that
will be replacing the old intent of the StitchedMap, I believe this new
version will be more lightweight and better handled. Prior the stitched
map have a large space consumption and a large load time. Instead this
new TextureArray comes with more efficiency in load and space
consumption at the cost of easily manageable textures due to the
required knowledge of all coordinates for the array. Another issue is
it's static nature and the inability for textures to be pushed together.